test: migrate cron-aggregate-insights to bdd
